requirements.txt
alembic==1.11.1
anyio==3.7.0 ; python_version >= '3.7'
asgiref==3.7.2 ; python_version >= '3.7'
bcrypt==4.0.1
certifi==2023.5.7 ; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
click==8.1.3 ; python_version >= '3.7'
cryptography==41.0.1
ecdsa==0.18.0
fastapi==0.68.2
fastapi-sqlalchemy==0.2.1
fastapi-sso==0.6.4
greenlet==2.0.2 ; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0 ; python_version >= '3.7'
httpcore==0.16.3 ; python_version >= '3.7'
httpx==0.23.3
idna==3.4 ; python_version >= '3.5'
iniconfig==2.0.0 ; python_version >= '3.7'
mako==1.2.4 ; python_version >= '3.7'
markupsafe==2.1.3 ; python_version >= '3.7'
oauthlib==3.2.2 ; python_version >= '3.6'
packaging==23.1 ; python_version >= '3.7'
passlib[bcrypt]==1.7.4
pluggy==1.0.0 ; python_version >= '3.6'
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.9
pyi18n-v2==1.2.1
pytest==7.3.2
pytest-mock==3.11.1
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pyyaml==6.0 ; python_version >= '3.6'
requests==2.31.0
rfc3986[idna2008]==1.5.0
rsa==4.9 ; python_version >= '3.6' and python_version < '4'
six==1.16.0
sniffio==1.3.0 ; python_version >= '3.7'
sqlalchemy==2.0.16
starlette==0.14.2 ; python_version >= '3.6'
typing-extensions==4.6.3 ; python_version >= '3.7'
urllib3==2.0.3 ; python_version >= '3.7'
uvicorn==0.15.0
void==0.1.3
